pkgdown/extra.css

Class Category

A S4 class for the Category level result of an Intensity analysis. Can be plotted with the plot method plot.

Details

The slots categoryData and categoryStationarity can receive tables for "Gain" or "Loss" in the following format:

Gain

categoryData: <tibble>. A table containing 6 columns:

Period: <fct>. The period [Yt, Yt+1].

To: <fct>. A LUC category j.

Interval: <int>. Duration of the period [Yt, Yt+1] in years.

GG_km2/GG_pixel: <dbl>/<int>. Area of gross gain of category j during [Yt, Yt+1].

Gtj: <dbl>. Annual intensity of gross gain of category j for time interval [Yt, Yt+1].

St: <dbl>. Annual intensity of change for time interval [Yt, Yt+1].

categoryStationarity: <tibble>. A table with the results of a stationarity test of the gain of the categories on the Category level, containing 5 columns:

To: <fct>. A category of interest j.

gain: <int>. Number of times a category had gains during all time intervals [Y1, YT].

N: <int>. Total number of evaluated time points (T).

Stationarity: <chr>. Active Gain or Dormant Gain.

Test: <chr>. Y if stationarity was detected and N if not.

Loss

categoryData: <tibble>. A table containing 6 columns:

Period: <fct>. The period [Yt, Yt+1].

From: <fct>. A LUC category i.

Interval: <int>. Duration of the period [Yt, Yt+1] in years.

GG_km2/GG_pixel: <dbl>/<int>. Area of gross loss of category i during [Yt, Yt+1].

Lti: <dbl>. Annual intensity of gross loss of category i for time interval [Yt, Yt+1].

STt: <dbl>. Annual intensity of change for time interval [Yt, Yt+1].

categoryStationarity: <tibble>. A table of stationarity test over the loss of the categories in the Category level, containing 5 columns:

From: <fct>. A category of interest i.

loss: <int>. Number of times a category had losses during all time intervals [Y1, YT].

N: <int>. Total number of evaluated time points (T).

Stationarity: <chr>. Active Loss or Dormant Loss.

Test: <chr>. Y if stationarity was detected and N if not.

Slots

lookupcolor

The colors (character vector) associated with the LUC legend items.

categoryData

tibble. A table of Category level's results (gain (Gtj) or loss (Lti) values).

categoryStationarity

tibble. A table containing results of a stationarity test. A change is considered stationary only if the intensities for all time intervals reside on one side of the uniform intensity, i.e are smaller or bigger than the uniform rate over the whole period.