Rasters of Land Use and Land Cover of the São Lourenço River Basin (2002 - 2014)

A multi-layer raster (RasterStack) with five (5) RasterLayer objects represinting the years 2002, 2008, 2010, 2012 and 2014 respectively. These Raster* have everyone 11 classes of use, these classes are:

Pixel Value	Legend	Class	Use	Details

2	Ap	Anthropic	Anthropic Use	Pastagem

3	FF	Natural	NA	Formações Florestais

4	SA	Natural	NA	Savana Arborizada (SA - Cerrado)

5	SG	Natural	NA	Savana Gramínea (SG - Campo)

7	aa	Anthropic	NA	Alteração Antrópica

8	SF	Natural	NA	Savana Florestada (SF - Cerradão)

9	Agua	Natural	NA	Rios, córregos, corixos, vazantes, baías e salinas

10	Iu	Anthropic	Anthropic Use	Influencia Urbana

11	Ac	Anthropic	Anthropic Use	Agricultura

12	R	Anthropic	Anthropic Use	Reflorestamento

13	Im	Anthropic	Anthropic Use	Degradação por Mineração

data(SaoLourencoBasin)

Format

An object of class "RasterBrick" with 5 layers:

landscape_2002: Land scape of the land use and land cover for the yeear 2002

landscape_2008: Land scape of the land use and land cover for the yeear 2008

landscape_2010: Land scape of the land use and land cover for the yeear 2010

landscape_2012: Land scape of the land use and land cover for the yeear 2012

landscape_2014: Land scape of the land use and land cover for the yeear 2014

Source

https://www.embrapa.br/pantanal/bacia-do-alto-paraguai

References

Instituto SOS Pantanal and WWF-Brasil (2015) (Monitoramento das alterações da cobertura vegetal e uso do Solo na Bacia do Alto Paraguai – Porção Brasileira – Período de Análise: 2012 a 2014. 4th edn. Brasilia.)

Examples

data(SaoLourencoBasin)
raster::plot(SaoLourencoBasin[[1]])

#> Error in .local(.Object, ...):