Contingence Table

Contingence Table

contingenceTable(input_raster, pixelresolution = 30)

Arguments

input_raster
list (of filenames or Raster* objects), RasterStack(brick) or RasterStack(stack)

pixelresolution
numeric. The pixel resolution in meter.

Value

A list that contains 5 objects.

lulc_Mulstistep: <tibble> The table of contingency between all step of time analysed, contains 8 columns:

Period: <chr> The period [Yt, Yt+1].

From: <dbl> a category i.

To: <dbl> a category j.

km2: <dbl> The quantity in kilometer that transit from the classes category i to category j in the period [Yt, Yt+1].

Interval: <dbl> The number of time point between the first and the last year of the period [Yt, Yt+1].

QtPixel: <int> The quantity of pixel that transit from the classes category i to category j in the period [Yt, Yt+1].

yearFrom: <chr> The year that the change come from [Yt]

yearTo: <chr> The year that the change go for [Yt+1]

lulc_Onestep:<tibble> The table of contingency between the first and the last year analysed [Yt, Yt+1], contains 8 columns like lulc_Mulstistep.

tb_legend: <tibble> A table of the pixel value his name and color, contains 3 columns:

classValue: <dbl> the pixel value of the classes of land use

className: <factor> The name or legend associate with a given pixel value, here as factor with a specific level for blablabla

color: <chr> The color associate with the given pixel value. Befor others step of analysis, one would want to change the className and color values. Some details.

the class name have to be in the same order of the classValue column, the levels the order that the graphs legend will have to be printed. Like:

        myobject$tb_legend$className <- factor(c("name1", "name2", "name3", "name4"),
                                               levels = c("name3", "name2", "name1", "name4"))

the color by the same order of the classValue comuln, it can be a colour name (eg. "black") or an HEX value (eg. #FFFFFF). Like:

       myobject$tb_legend$color <- c("#CDB79E", "red", "#66CD00", "yellow")

totalArea: <tibble> A table of two columns with the total area of study, contains 2 columns:

area_km2: <numeric> The total area of study in square meter.

QtPixel: <numeric> The total area of study in quantity of pixel

totalInterval: <numeric> The value represinting the whole time points interval of time analysed [Yt, Yt+1], in years.

Examples

contingenceTable(demo_landscape(year = 2000:2005, res = 1), pixelresolution = 1)

#> $lulc_Multistep
#> # A tibble: 125 x 8
#>    Period     From    To      km2 QtPixel Interval yearFrom yearTo
#>    <chr>     <int> <int>    <dbl>   <int>    <int>    <int>  <int>
#>  1 2000-2001     1     1 0.000429     429        1     2000   2001
#>  2 2000-2001     1     2 0.000412     412        1     2000   2001
#>  3 2000-2001     1     3 0.000366     366        1     2000   2001
#>  4 2000-2001     1     4 0.000379     379        1     2000   2001
#>  5 2000-2001     1     5 0.000382     382        1     2000   2001
#>  6 2000-2001     2     1 0.0004       400        1     2000   2001
#>  7 2000-2001     2     2 0.000385     385        1     2000   2001
#>  8 2000-2001     2     3 0.000408     408        1     2000   2001
#>  9 2000-2001     2     4 0.000413     413        1     2000   2001
#> 10 2000-2001     2     5 0.00038      380        1     2000   2001
#> # … with 115 more rows
#> 
#> $lulc_Onestep
#> # A tibble: 25 x 8
#>    Period     From    To      km2 QtPixel Interval yearFrom yearTo
#>    <chr>     <int> <int>    <dbl>   <int>    <int>    <int>  <int>
#>  1 2000-2005     1     1 0.000395     395        5     2000   2005
#>  2 2000-2005     1     2 0.000373     373        5     2000   2005
#>  3 2000-2005     1     3 0.000418     418        5     2000   2005
#>  4 2000-2005     1     4 0.000383     383        5     2000   2005
#>  5 2000-2005     1     5 0.000399     399        5     2000   2005
#>  6 2000-2005     2     1 0.000424     424        5     2000   2005
#>  7 2000-2005     2     2 0.000366     366        5     2000   2005
#>  8 2000-2005     2     3 0.000396     396        5     2000   2005
#>  9 2000-2005     2     4 0.000404     404        5     2000   2005
#> 10 2000-2005     2     5 0.000396     396        5     2000   2005
#> # … with 15 more rows
#> 
#> $tb_legend
#> # A tibble: 5 x 3
#>   classValue className color  
#>        <int> <fct>     <chr>  
#> 1          1 EOD       #002F70
#> 2          2 JMX       #879FDB
#> 3          3 GJV       #F6F6F6
#> 4          4 AJQ       #DA8A8B
#> 5          5 ZIS       #5F1415
#> 
#> $totalArea
#> # A tibble: 1 x 2
#>   area_km2 QtPixel
#>      <dbl>   <int>
#> 1     0.01   10000
#> 
#> $totalInterval
#> [1] 5
#>