Create Raster with Random pixel Value

This function create a raster serie with some setup like the layer name and the sample value for the lulc

demo_landscape(year, nrows = 100, ncols = 100, res = 1, xmn = 0,
  xmx = 100, ymn = 0, ymx = 100, crs = NA, class = 1:5,
  prob = NULL, sim.autocorrelate = FALSE, bin.interval = TRUE,
  psill = 0.025, model = "Exp", range = 1000,
  variance = c("identity", "mu", "mu(1-mu)")[1])

Arguments

year
A list of year, first and last included

nrows
nrows of the raster to be created

ncols
ncols of the raster to be created

res
the resolution of the raster to be created

xmn
x minimun extent

xmx
x maximum extent

ymn
y minimum extent

ymx
y maximum extent

crs
the coordinate referencing system

class
the raster classes

prob
the probability of occorence for the class list

sim.autocorrelate
logical This argument define if the simulation will be a autocorrelate in space or pixel of independent values.

bin.interval
logical This argument define if the categorisation of the value will be via igual interval or igual number see `ggplot2`

psill
sill see variogram

model
a model see variogram

range
a range of the variability

variance
character; variance function to transform to non-stationary covariances; "identity" does not transform, other options are "mu" (Poisson) and "mu(1-mu)" (binomial)

Value

list

Examples

demo_landscape(2000:2005, res = 1)

#> [[1]]
#> class      : RasterLayer 
#> dimensions : 100, 100, 10000  (nrow, ncol, ncell)
#> resolution : 1, 1  (x, y)
#> extent     : 0, 100, 0, 100  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : landscape_2000 
#> values     : 1, 5  (min, max)
#> 
#> 
#> [[2]]
#> class      : RasterLayer 
#> dimensions : 100, 100, 10000  (nrow, ncol, ncell)
#> resolution : 1, 1  (x, y)
#> extent     : 0, 100, 0, 100  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : landscape_2001 
#> values     : 1, 5  (min, max)
#> 
#> 
#> [[3]]
#> class      : RasterLayer 
#> dimensions : 100, 100, 10000  (nrow, ncol, ncell)
#> resolution : 1, 1  (x, y)
#> extent     : 0, 100, 0, 100  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : landscape_2002 
#> values     : 1, 5  (min, max)
#> 
#> 
#> [[4]]
#> class      : RasterLayer 
#> dimensions : 100, 100, 10000  (nrow, ncol, ncell)
#> resolution : 1, 1  (x, y)
#> extent     : 0, 100, 0, 100  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : landscape_2003 
#> values     : 1, 5  (min, max)
#> 
#> 
#> [[5]]
#> class      : RasterLayer 
#> dimensions : 100, 100, 10000  (nrow, ncol, ncell)
#> resolution : 1, 1  (x, y)
#> extent     : 0, 100, 0, 100  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : landscape_2004 
#> values     : 1, 5  (min, max)
#> 
#> 
#> [[6]]
#> class      : RasterLayer 
#> dimensions : 100, 100, 10000  (nrow, ncol, ncell)
#> resolution : 1, 1  (x, y)
#> extent     : 0, 100, 0, 100  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : landscape_2005 
#> values     : 1, 5  (min, max)
#> 
#>