Perform the intensity analysis from a cross-tabulation matrices

This function inplement the Intensity Analysis, a quantitative method to analyze maps of land categories from several points in time for a single site by considering cross-tabulation matrices, where one matrix summarizes the change in each time interval.

intensityanalysis(dataset, class_n, class_m, area_km2 = TRUE)

Arguments

dataset
The list of object from contingenceTable.

class_n
character. the gaining category in the transition of interest n.

class_m
character. the losing category in the transition of interest m.

area_km2
logical. Whether the change would be computed in km2 or quantity of pixel unit.

Value

Intensity objects including the levels tables and strationarity test tables

Details

There are three levels of analysis, starting from general to more detailed levels, where each level exposes different types of information given the previous level of analysis (Aldwaik and Pontius, 2012, 2013).

The interval level examines how the size and speed of change vary across time intervals.

The category level examines how the size and intensity of gross losses and gross gains in each category vary across categories for each time interval.

The transition level examines how the size and intensity of a category’s transitions vary across the other categories that are available for that transition.

At each level, the method tests for stationarity of patterns across time intervals.

The function return a list with 10 objects:

lulc_table: A tibble of contingency between all step of time analysed, contains 6 columns:

Period: <fct>. The period of change in the format first year - last year.

From: <fct>. The class in the first year.

To: <fct>. The class in the last year.

km2: dbl. The quantity in kilometer that transit from the classes From to To in the period.

interval: dbl. The number of time point between the first and the last year of the period.

QtPixel: int. The quantity in number of pixel that transits from the classes From to To in the period.

lv1_tbl: An "IntensityL01" object containing 1 table with 4 columns:

Period: <fct>. The period [Yt, Yt+1].

PercentChange: <dbl>. Interval Change Area (percent of map).

St: <dbl>. annual intensity of change for time interval [Yt, Yt+1].

U: <dbl>. value of uniform line for time intensity analysis.

lv2_gain: An "IntensityL02" object containing a slot with lookup table af the colors associate with the land classes and a table with 6 columns:

Period: <fct>. The period [Yt, Yt+1].

To: <fct>. a category j.

interval: <dbl>. duration of the period [Yt, Yt+1].

GG_km2/GG_pixel: <dbl>/<int>. area of gross gain of category j during [Yt, Yt+1].

Gtj: <dbl>. annual intensity of gross gain of category j for time interval [Yt, Yt+1].

St: <dbl>. annual intensity of change for time interval [Yt, Yt+1].

lv2_loss: An "IntensityL02" object containing a slot with lookup table af the colors associate with the land classes and a table with 6 columns:

Period: <fct>. The period [Yt, Yt+1].

From: <fct>. a category i.

interval: <dbl>. duration of the period [Yt, Yt+1].

GG_km2/GG_pixel: <dbl>/<int>. area of gross loss of category j during [Yt, Yt+1].

Lti: <dbl>. annual intensity of gross loss of category i for time interval [Yt, Yt+1].

STt: <dbl>. annual intensity of change for time interval [Yt, Yt+1].

gain_n: An "IntensityL03" object containing a slot with lookup table af the colors associate with the land classes and a table with 7 columns:

Period: <fct>. The period [Yt, Yt+1].

From: <fct>. a category i.

To: <fct>. the gaining category in the transition of interest (n).

interval: <dbl>. duration of the period [Yt, Yt+1].

T_i2n_km2/T_i2n_pixel: <dbl>. transition from category i to category n during time interval [Yt, Yt+1] where i is not equal to n.

Rtin: <dbl>. annual intensity of transition from category i to category n during time interval [Yt, Yt+1] where i is not equal to n.

Wtn: <dbl>. value of uniform intensity of transition to category n from all non-n categories at time Yt during time interval [Yt, Yt+1].

loss_m: An "IntensityL03" object containing a slot with lookup table af the colors associate with the land classes and a table with 7 columns:

Period: <fct>. The period [Yt, Yt+1].

To: <fct>. a category j.

From: <fct>. the losing category in the transition of interest (m).

interval: <dbl>. duration of the period [Yt, Yt+1].

T_m2j_km2/T_m2j_pixel: <dbl>. a transition from category m to category j during time interval [Yt, Yt+1] where j is not equal to m.

Qtmj: <dbl>. annual intensity of transition from category m to category j during time interval [Yt, Yt+1] where j is not equal to m.

Vtm: <dbl>. value of uniform intensity of transition from category m to all non-m categories at time Yt+1 during time interval [Yt, Yt+1].

st_lv2_gain: A tibble of stationarity test in level 02 gain containing 5 columns:

To: <fct>. acategory j.

gain: <int>. Number of time the class loss during time interval [Yt, Yt+1].

N: <int>. Number total of time of trasition to be considered as stationary (T).

Stationarity: <chr>. Active Gain or Dormant Gain.

Test: <chr>. Y for stationarity detected and N when not.

st_lv2_loss: A tibble of stationarity test in level 02 loss containing 5 columns:

From: <fct>. acategory i.

loss: <int>. Number of time the class loss during time interval [Yt, Yt+1].

N: <int>. Number total of time of trasition to be considered as stationary (T).

Stationarity: <chr>. Active Loss or Dormant Loss.

Test: <chr>. Y for stationarity detected and N when not.

st_gain_n: A tibble of stationarity test in level 03 gain on class n containing 5 columns:

From: <fct>. the losing category in the transition of interest to the n category.

loss: <int>. Number of time the class loss to the n class.

N: <int>. Number total of time of trasition to be considered as stationary (T).

Stationarity: <chr>. targeted by or avoided by by the n class.

Test: <chr>. Y for stationarity detected and N when not.

st_loss_m: A tibble of stationarity test in level 03 loss on class m containing 5 columns:

To: <fct>. the gaining category in the transition of interest from the m category.

gain: <int>. Number of time the class gain from the m class.

N: <int>. Number total of time of trasition to be considered as stationary (T).

Stationarity: <chr>. targeted or avoided by the m class.

Test: <chr>. Y for stationarity detected and N when not.

References

Aldwaik, S. Z. and Pontius, R. G. (2012) ‘Intensity analysis to unify measurements of size and stationarity of land changes by interval, category, and transition’, Landscape and Urban Planning. Elsevier B.V., 106(1), pp. 103–114. doi: 10.1016/j.landurbplan.2012.02.010 .

Aldwaik, S. Z. and Pontius, R. G. (2013) ‘Map errors that could account for deviations from a uniform intensity of land change’, International Journal of Geographical Information Science. Taylor & Francis, 27(9), pp. 1717–1739. doi: 10.1080/13658816.2013.787618 .

Examples

intensityanalysis(dataset = SL_2002_2014, class_n = "Ap", class_m = "SG", area_km2 = TRUE)

#> $lulc_table
#> # A tibble: 130 x 6
#>    Period    From  To         km2 interval QtPixel
#>    <fct>     <fct> <fct>    <dbl>    <dbl>   <int>
#>  1 2002-2008 Ap    Ap    6543.           6 7269961
#>  2 2002-2008 Ap    Iu       1.56         6    1736
#>  3 2002-2008 Ap    Ac      55.2          6   61320
#>  4 2002-2008 Ap    R       23.9          6   26609
#>  5 2002-2008 FF    Ap      37.5          6   41649
#>  6 2002-2008 FF    FF    2133.           6 2370190
#>  7 2002-2008 FF    aa     155.           6  172718
#>  8 2002-2008 FF    Ac       7.48         6    8307
#>  9 2002-2008 FF    R        0.356        6     395
#> 10 2002-2008 FF    Im       0.081        6      90
#> # ... with 120 more rows
#> 
#> $lv1_tbl
#> An object of class "IntensityL01"
#> Slot "level":
#> [1] "Interval"
#> 
#> Slot "tabela":
#> # A tibble: 4 x 4
#> # Groups:   Period [4]
#>   Period    PercentChange    St     U
#>   <fct>             <dbl> <dbl> <dbl>
#> 1 2012-2014         3.32  1.66   1.13
#> 2 2010-2012         4.23  2.12   1.13
#> 3 2008-2010         0.880 0.440  1.13
#> 4 2002-2008         5.18  0.864  1.13
#> 
#> 
#> $lv2_gain
#> An object of class "IntensityL02"
#> Slot "level":
#> [1] "Category"
#> 
#> Slot "color":
#>        Ap        FF        SA        SG        aa        SF      Agua        Iu 
#> "#CDB79E" "#006400" "#66CD00" "#7FFF00" "#EE2C2C" "#458B00" "#0000CD" "#6B6B6B" 
#>        Ac         R        Im 
#> "#CD950C" "#6B8E23" "#CD00CD" 
#> 
#> Slot "tabela":
#> # A tibble: 23 x 6
#> # Groups:   Period, To [23]
#>    Period    To    interval  GG_km2   Gtj    St
#>    <fct>     <fct>    <dbl>   <dbl> <dbl> <dbl>
#>  1 2012-2014 aa           2  14.9   0.510  1.66
#>  2 2012-2014 Ap           2 612.    3.92   1.66
#>  3 2012-2014 Ac           2 110.    1.14   1.66
#>  4 2012-2014 Im           2   0.195 0.337  1.66
#>  5 2012-2014 Iu           2   6.79  2.67   1.66
#>  6 2010-2012 aa           2  47.0   1.18   2.12
#>  7 2010-2012 Ap           2 707.    4.84   2.12
#>  8 2010-2012 Ac           2 189.    2.00   2.12
#>  9 2010-2012 Iu           2   1.90  0.792  2.12
#> 10 2010-2012 R            2   2.76  0.951  2.12
#> # ... with 13 more rows
#> 
#> 
#> $lv2_loss
#> An object of class "IntensityL02"
#> Slot "level":
#> [1] "Category"
#> 
#> Slot "color":
#>        Ap        FF        SA        SG        aa        SF      Agua        Iu 
#> "#CDB79E" "#006400" "#66CD00" "#7FFF00" "#EE2C2C" "#458B00" "#0000CD" "#6B6B6B" 
#>        Ac         R        Im 
#> "#CD950C" "#6B8E23" "#CD00CD" 
#> 
#> Slot "tabela":
#> # A tibble: 29 x 6
#> # Groups:   Period, From [29]
#>    Period    From  interval GL_km2     Lti    St
#>    <fct>     <fct>    <dbl>  <dbl>   <dbl> <dbl>
#>  1 2012-2014 FF           2  15.4   0.365   1.66
#>  2 2012-2014 SF           2   3.49  0.224   1.66
#>  3 2012-2014 SA           2  25.3   0.845   1.66
#>  4 2012-2014 SG           2  46.9   0.634   1.66
#>  5 2012-2014 aa           2 541.   13.6     1.66
#>  6 2012-2014 Ap           2 111.    0.759   1.66
#>  7 2012-2014 Ac           2   1.26  0.0133  1.66
#>  8 2010-2012 FF           2  11.1   0.263   2.12
#>  9 2010-2012 SF           2   4.44  0.283   2.12
#> 10 2010-2012 SA           2  29.7   0.974   2.12
#> # ... with 19 more rows
#> 
#> 
#> $gain_n
#> An object of class "IntensityL03"
#> Slot "level":
#> [1] "Transition"
#> 
#> Slot "color":
#>        Ap        FF        SA        SG        aa        SF      Agua        Iu 
#> "#CDB79E" "#006400" "#66CD00" "#7FFF00" "#EE2C2C" "#458B00" "#0000CD" "#6B6B6B" 
#>        Ac         R        Im 
#> "#CD950C" "#6B8E23" "#CD00CD" 
#> 
#> Slot "tabela":
#> # A tibble: 20 x 7
#> # Groups:   Period, From [20]
#>    Period    From  To    interval T_i2n_km2    Rtin    Wtn
#>    <fct>     <fct> <fct>    <dbl>     <dbl>   <dbl>  <dbl>
#>  1 2012-2014 FF    Ap           2    12.9    0.307  2.03  
#>  2 2012-2014 SF    Ap           2     2.38   0.152  2.03  
#>  3 2012-2014 SA    Ap           2    18.1    0.605  2.03  
#>  4 2012-2014 SG    Ap           2    42.1    0.569  2.03  
#>  5 2012-2014 aa    Ap           2   537.    13.5    2.03  
#>  6 2010-2012 FF    Ap           2     3.63   0.0856 2.26  
#>  7 2010-2012 SF    Ap           2     1.08   0.0685 2.26  
#>  8 2010-2012 SA    Ap           2    12.3    0.403  2.26  
#>  9 2010-2012 SG    Ap           2     8.53   0.114  2.26  
#> 10 2010-2012 aa    Ap           2   682.    13.0    2.26  
#> 11 2008-2010 FF    Ap           2     0.968  0.0227 0.0610
#> 12 2008-2010 SF    Ap           2     1.10   0.0694 0.0610
#> 13 2008-2010 SA    Ap           2     0.971  0.0314 0.0610
#> 14 2008-2010 SG    Ap           2     3.94   0.0524 0.0610
#> 15 2008-2010 aa    Ap           2    12.0    0.233  0.0610
#> 16 2002-2008 FF    Ap           6    37.5    0.268  0.323 
#> 17 2002-2008 SF    Ap           6    24.8    0.453  0.323 
#> 18 2002-2008 SA    Ap           6    65.8    0.608  0.323 
#> 19 2002-2008 SG    Ap           6    48.3    0.198  0.323 
#> 20 2002-2008 aa    Ap           6   130.     1.02   0.323 
#> 
#> 
#> $loss_m
#> An object of class "IntensityL03"
#> Slot "level":
#> [1] "Transition"
#> 
#> Slot "color":
#>        Ap        FF        SA        SG        aa        SF      Agua        Iu 
#> "#CDB79E" "#006400" "#66CD00" "#7FFF00" "#EE2C2C" "#458B00" "#0000CD" "#6B6B6B" 
#>        Ac         R        Im 
#> "#CD950C" "#6B8E23" "#CD00CD" 
#> 
#> Slot "tabela":
#> # A tibble: 14 x 7
#> # Groups:   Period, To [14]
#>    Period    To    From  interval T_m2j_km2     Qtmj    Vtm
#>    <fct>     <fct> <fct>    <dbl>     <dbl>    <dbl>  <dbl>
#>  1 2012-2014 aa    SG           2    4.76   0.163    0.125 
#>  2 2012-2014 Ap    SG           2   42.1    0.270    0.125 
#>  3 2012-2014 Ac    SG           2    0.0621 0.000642 0.125 
#>  4 2010-2012 aa    SG           2   18.9    0.475    0.0749
#>  5 2010-2012 Ap    SG           2    8.53   0.0584   0.0749
#>  6 2010-2012 Ac    SG           2    0.632  0.00669  0.0749
#>  7 2008-2010 aa    SG           2   30.6    0.584    0.0929
#>  8 2008-2010 Ap    SG           2    3.94   0.0290   0.0929
#>  9 2008-2010 Ac    SG           2    0.0873 0.000962 0.0929
#> 10 2008-2010 Iu    SG           2    0.0504 0.0213   0.0929
#> 11 2002-2008 aa    SG           6  244.     1.58     0.273 
#> 12 2002-2008 Ap    SG           6   48.3    0.118    0.273 
#> 13 2002-2008 Ac    SG           6   13.1    0.0489   0.273 
#> 14 2002-2008 R     SG           6    0.0999 0.0129   0.273 
#> 
#> 
#> $st_lv2_gain
#> # A tibble: 12 x 5
#>    To     Gain     N Stationarity Test 
#>    <fct> <int> <int> <chr>        <chr>
#>  1 aa        2     4 Active Gain  N    
#>  2 Ap        2     4 Active Gain  N    
#>  3 Ac        1     4 Active Gain  N    
#>  4 Iu        2     4 Active Gain  N    
#>  5 Agua      1     4 Active Gain  N    
#>  6 R         2     4 Active Gain  N    
#>  7 aa        2     4 Dormant Gain N    
#>  8 Ap        2     4 Dormant Gain N    
#>  9 Ac        3     4 Dormant Gain N    
#> 10 Im        3     4 Dormant Gain N    
#> 11 Iu        2     4 Dormant Gain N    
#> 12 R         1     4 Dormant Gain N    
#> 
#> $st_lv2_loss
#> # A tibble: 14 x 5
#>    From   Loss     N Stationarity Test 
#>    <fct> <int> <int> <chr>        <chr>
#>  1 FF        1     4 Active Loss  N    
#>  2 SF        2     4 Active Loss  N    
#>  3 SA        2     4 Active Loss  N    
#>  4 SG        2     4 Active Loss  N    
#>  5 aa        3     4 Active Loss  N    
#>  6 Ap        1     4 Active Loss  N    
#>  7 R         1     4 Active Loss  N    
#>  8 FF        3     4 Dormant Loss N    
#>  9 SF        2     4 Dormant Loss N    
#> 10 SA        2     4 Dormant Loss N    
#> 11 SG        2     4 Dormant Loss N    
#> 12 aa        1     4 Dormant Loss N    
#> 13 Ap        3     4 Dormant Loss N    
#> 14 Ac        4     4 Dormant Loss Y    
#> 
#> $st_gain_n
#> # A tibble: 7 x 5
#>   From   Loss     N Stationarity   Test 
#>   <fct> <int> <int> <chr>          <chr>
#> 1 SF        2     4 targeted by Ap N    
#> 2 SA        1     4 targeted by Ap N    
#> 3 aa        4     4 targeted by Ap Y    
#> 4 FF        4     4 avoided by Ap  Y    
#> 5 SF        2     4 avoided by Ap  N    
#> 6 SA        3     4 avoided by Ap  N    
#> 7 SG        4     4 avoided by Ap  Y    
#> 
#> $st_loss_m
#> # A tibble: 6 x 5
#>   To     Gain     N Stationarity Test 
#>   <fct> <int> <int> <chr>        <chr>
#> 1 aa        4     4 targeted SG  Y    
#> 2 Ap        1     4 targeted SG  N    
#> 3 Ap        3     4 avoided SG   N    
#> 4 Ac        4     4 avoided SG   Y    
#> 5 Iu        1     4 avoided SG   N    
#> 6 R         1     4 avoided SG   N    
#>