Tables of Land Use and Land Cover of the São Lourenço River Basin (2002 - 2014)

A list containing five objects from contingenceTable funtion with the SaoLourencoBasin as input (SL_2002_2014 <- contingenceTable(input_raster = SaoLourencoBasin, pixelresolution = 30)).

data(SL_2002_2014)

Format

A data list with 5 objects:

lulc_Multistep

<tibble> A table of contingency between all step of time analysed, contains 8 columns:

Period: <chr> period between the years analyzed

From: <int> The pixel value from

To: <int> The pixel value to

km2: <dbl> The area of change in squared kilometer

QtPixel: <int> The change in quatity of pixel

Interval: <int> The interval in years between the period analyzed

yearFrom: <int> The year that the change come from

yearTo: <int> The year that the change go for

lulc_Onstep

<tibble> A table of contingensy of the first and last year analysed with the same columns as lulc_Mulstistep tibble

tb_legend

<tibble> A table of the pixel value, their names and colors, contains 3 columns:

classValue: <int> the pixel value of the classes of land use

className: <fct> The name or legend associated with a given pixel value

color: <chr> The color associate with the given pixel value

totalArea

<tibble> A table of two columns with the total area of study, contains 2 columns:

area_km2: <dbl> The total area of study in km2

QtPixel: <int> The total area of study in quantity of pixel

totalInterval

<int> A number represinting the whole interval of time analysed

Source

https://www.embrapa.br/pantanal/bacia-do-alto-paraguai