Tables of land use and cover (LUC) in the São Lourenço River Basin (2002 - 2014)

A list containing five objects created by the contingencyTable function with SaoLourencoBasin as input (SL_2002_2014 <- contingenceTable(input_raster = SaoLourencoBasin, pixelresolution = 30)).

data(SL_2002_2014)

Format

A data list with 5 objects:

lulc_Multistep

<tibble> Contingency table for all analysed time steps, containing 8 columns:

Period: <chr> The period [Yt, Yt+1].

From: <int> numerical code of a LUC category i.

To: <int> numerical code of a LUC category j.

km2: <dbl> Area in square kilometers that transited from the category i to category j in the period from Yt to Yt+1.

QtPixel: <int> Pixel count that transited from the categories i to category j in the period from Yt to Yt+1.

Interval: <int> Interval of years between the first and the last year of the period [Yt, Yt+1].

yearFrom: <int> The year that the change comes from [Yt]

yearTo: <int> The year that the change goes for [Yt+1]

lulc_Onstep

<tibble> Contingency table for the entire analysed period [Yt1, YT], containing 8 columns identical with lulc_Mulstistep

.
tb_legend

<tibble> A table of the pixel value, his name and color containing 3 columns:

categoryValue: <int> the pixel value of the LUC category.

categoryName: <fct> randomly created string associated with a given pixel value of a LUC category.

color: <chr> random color associated with the given pixel value of a LUC category.

totalArea

<tibble> A table with the total area of the study area containing 2 columns:

area_km2: <dbl> The total area in square kilometers.

QtPixel: <int> The total area in pixel counts

.
totalInterval

<int> Total interval of the analysed time series in years

.
Source

https://www.embrapa.br/pantanal/bacia-do-alto-paraguai