Class Transition

An S4 class for the Transition level analysis result, the objects of this class can be plot with the plot method plot

Details

The slots transitionData and transitionStationarity can receive tables for "Gain of class n" or "Loss of class m" in the following format:

Gain of class n:

transitionData: <tibble>. A table with 7 columns:

Period: <fct>. The period [Yt, Yt+1].

From: <fct>. a category i.

To: <fct>. the gaining category in the transition of interest (n).

Interval: <int>. duration of the period [Yt, Yt+1].

T_i2n_km2/T_i2n_pixel: <dbl>. transition from category i to category n during time interval [Yt, Yt+1] where i is not equal to n.

Rtin: <dbl>. annual intensity of transition from category i to category n during time interval [Yt, Yt+1] where i is not equal to n.

Wtn: <dbl>. value of uniform intensity of transition to category n from all non-n categories at time Yt during time interval [Yt, Yt+1].

transitionStationarity: <tibble>. A table of stationarity test over the gain on class n containing 5 columns:

From: <fct>. the losing category in the transition of interest to the n category.

loss: <int>. Number of time the class loss to the n class.

N: <int>. Number total of time of trasition to be considered as stationary (T).

Stationarity: <chr>. targeted by or avoided by by the n class.

Test: <chr>. Y for stationarity detected and N when not.

Loss of class m:

transitionData: <tibble>. A table with 7 columns:

Period: <fct>. The period [Yt, Yt+1].

To: <fct>. a category j.

From: <fct>. the losing category in the transition of interest (m).

Interval: <dbl>. duration of the period [Yt, Yt+1].

T_m2j_km2/T_m2j_pixel: <dbl>. a transition from category m to category j during time interval [Yt, Yt+1] where j is not equal to m.

Qtmj: <dbl>. annual intensity of transition from category m to category j during time interval [Yt, Yt+1] where j is not equal to m.

Vtm: <dbl>. value of uniform intensity of transition from category m to all non-m categories at time Yt+1 during time interval [Yt, Yt+1].

transitionStationarity: <tibble>. A table of stationarity test over the loss of class m containing 5 columns:

To: <fct>. the gaining category in the transition of interest from the m category.

gain: <int>. Number of time the class gain from the m class.

N: <int>. Number total of time of trasition to be considered as stationary (T).

Stationarity: <chr>. targeted or avoided by the m class.

Test: <chr>. Y for stationarity detected and N when not.

Slots

lookupcolor

character; The colors associate with the data with the legends as name’s attribute

transitionData

tibble; A table of Transition level's result (gain n (Rtin & Wtn) or loss m (Qtmj & Vtm) values)

transitionStationarity

tibble; A table containing a stationarity test. A change is take as stationary only whether the intensities for all time intervals reside on one side of the uniform intensity, i.e, smaller or bigger than the uniform rate over the whole period.