Accumulates changes in a LULC raster time series

This function calculates the number of times a pixel has changed during the analysed period. It returns a raster with the number of changes as pixel value and a table containing the areal percentage of every pixel value (number of changes).

acc_changes(path)

Arguments

path
list. List of filenames, list of Raster* objects, RasterStack(brick) or RasterStack(stack)

Value

Two objects, a raster layer and a table.

Examples

acc_changes(SaoLourencoBasin)

#> [[1]]
#> class      : RasterLayer 
#> dimensions : 6372, 6546, 41711112  (nrow, ncol, ncell)
#> resolution : 30, 30  (x, y)
#> extent     : 654007.5, 850387.5, 8099064, 8290224  (xmin, xmax, ymin, ymax)
#> crs        : +proj=utm +zone=21 +south +ellps=GRS80 +units=m +no_defs 
#> source     : C:/Users/pg_reginal/AppData/Local/Temp/Rtmp0UW8Jf/raster/r_tmp_2020-01-29_131401_3848_64535.grd 
#> names      : layer 
#> values     : 0, 2  (min, max)
#> 
#> 
#> [[2]]
#> # A tibble: 3 x 3
#>   PxValue       Qt Percent
#>     <int>    <int>   <dbl>
#> 1       0 21819779   87.6 
#> 2       1  2787995   11.2 
#> 3       2   301086    1.21
#>