Accumulate changes in a LULC time serie raster

This fucntion return two maps, the first containing the sequence of change as pixel value (answering: in what point of time a pixel has changes?) and second the quantity time of change as pixel value(answering: how many times a pixel has change). These maps can be plotted with the plot function method of the raster package or with the `tmap` package.

acc_changes(path)

Arguments

path
list (of filenames or Raster* objects), RasterStack(brick) or RasterStack(stack)

Value

Two maps, the first containing the sequence of change as pixel value and second the quantity time of change as pixel value.

Examples

test <- demo_landscape(2000:2005, res = 1, prob = c(0.05, 0.3, 0.05, 0.4, 0.2))
acc_changes(test)

#> [[1]]
#> class      : RasterLayer 
#> dimensions : 100, 100, 10000  (nrow, ncol, ncell)
#> resolution : 1, 1  (x, y)
#> extent     : 0, 100, 0, 100  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 5  (min, max)
#> 
#> 
#> [[2]]
#> # A tibble: 6 x 3
#>   PxValue    Qt Percent
#>     <int> <int>   <dbl>
#> 1       0    38    0.38
#> 2       1   315    3.15
#> 3       2  1268   12.7 
#> 4       3  2879   28.8 
#> 5       4  3562   35.6 
#> 6       5  1938   19.4 
#> 

acc_changes(raster::stack(test))

#> [[1]]
#> class      : RasterLayer 
#> dimensions : 100, 100, 10000  (nrow, ncol, ncell)
#> resolution : 1, 1  (x, y)
#> extent     : 0, 100, 0, 100  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 5  (min, max)
#> 
#> 
#> [[2]]
#> # A tibble: 6 x 3
#>   PxValue    Qt Percent
#>     <int> <int>   <dbl>
#> 1       0    38    0.38
#> 2       1   315    3.15
#> 3       2  1268   12.7 
#> 4       3  2879   28.8 
#> 5       4  3562   35.6 
#> 6       5  1938   19.4 
#> 

acc_changes(raster::brick(test))

#> [[1]]
#> class      : RasterLayer 
#> dimensions : 100, 100, 10000  (nrow, ncol, ncell)
#> resolution : 1, 1  (x, y)
#> extent     : 0, 100, 0, 100  (xmin, xmax, ymin, ymax)
#> crs        : NA 
#> source     : memory
#> names      : layer 
#> values     : 0, 5  (min, max)
#> 
#> 
#> [[2]]
#> # A tibble: 6 x 3
#>   PxValue    Qt Percent
#>     <int> <int>   <dbl>
#> 1       0    38    0.38
#> 2       1   315    3.15
#> 3       2  1268   12.7 
#> 4       3  2879   28.8 
#> 5       4  3562   35.6 
#> 6       5  1938   19.4 
#>