Perform the intensity analysis from a cross-tabulation matrices

This function inplement the Intensity Analysis, a quantitative method to analyze maps of land categories from several points in time for a single site by considering cross-tabulation matrices, where one matrix summarizes the change in each time interval.

intensityAnalysis(dataset, class_n, class_m, area_km2 = TRUE)

Arguments

dataset
list. The result object from contingenceTable.

class_n
character. the gaining category in the transition of interest n.

class_m
character. the losing category in the transition of interest m.

area_km2
logical. Whether the change would be computed in km2 or quantity of pixel unit.

Value

Intensity objects including the levels tables and strationarity test tables

Details

There are three levels of analysis, starting from general to more detailed levels, where each level exposes different types of information given the previous level of analysis (Aldwaik and Pontius, 2012, 2013).

The interval level examines how the size and speed of change vary across time intervals.

The category level examines how the size and intensity of gross losses and gross gains in each category vary across categories for each time interval.

The transition level examines how the size and intensity of a category’s transitions vary across the other categories that are available for that transition.

At each level, the method tests for stationarity of patterns across time intervals.

The function return a list with 6 objects:

lulc_table: tibble. A table of contingency between all step of time analysed, containing 6 columns:

Period: <fct>. The period of change in the format first year - last year.

From: <fct>. The class in the first year.

To: <fct>. The class in the last year.

km2: <dbl>. The quantity in kilometer that transit from the classes From to To in the period.

QtPixel: <int>. The quantity in number of pixel that transits from the classes From to To in the period.

Interval: <int>. The number of time point between the first and the last year of the period.

lv1_tbl: An Interval object (containing the St and U values).

category_lvlGain: A Category object (gain with the Gtj values).

category_lvlLoss: A Category object (loss with the Lti values).

transition_lvlGain_n: A Transition object (gain in class n with Rtin and Wtn values).

transition_lvlLoss_m: A Transition object (loss in class m with Qtmj and Vtm values).

References

Aldwaik, S. Z. and Pontius, R. G. (2012) ‘Intensity analysis to unify measurements of size and stationarity of land changes by interval, category, and transition’, Landscape and Urban Planning. Elsevier B.V., 106(1), pp. 103–114. doi: 10.1016/j.landurbplan.2012.02.010 .

Aldwaik, S. Z. and Pontius, R. G. (2013) ‘Map errors that could account for deviations from a uniform intensity of land change’, International Journal of Geographical Information Science. Taylor & Francis, 27(9), pp. 1717–1739. doi: 10.1080/13658816.2013.787618 .

Examples

intensityAnalysis(dataset = SL_2002_2014, class_n = "Ap", class_m = "SG", area_km2 = TRUE)

#> Warning: Factor `Period` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `From` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `Period` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `From` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `Period` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `From` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `Period` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `To` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `Period` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `To` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `Period` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `To` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `Period` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `From` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `Period` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `To` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `Period` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `To` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `Period` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `To` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `Period` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `From` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `Period` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `From` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `Period` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `From` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `Period` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `To` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `From` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `From` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `To` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `To` contains implicit NA, consider using `forcats::fct_explicit_na`

#> $lulc_table
#> # A tibble: 130 x 6
#>    Period    From  To         km2 QtPixel Interval
#>    <fct>     <fct> <fct>    <dbl>   <int>    <int>
#>  1 2002-2008 LEJ   LEJ   6543.    7269961        6
#>  2 2002-2008 LEJ   WLE      1.56     1736        6
#>  3 2002-2008 LEJ   WNH     55.2     61320        6
#>  4 2002-2008 LEJ   OQL     23.9     26609        6
#>  5 2002-2008 GXL   LEJ     37.5     41649        6
#>  6 2002-2008 GXL   GXL   2133.    2370190        6
#>  7 2002-2008 GXL   PZV    155.     172718        6
#>  8 2002-2008 GXL   WNH      7.48     8307        6
#>  9 2002-2008 GXL   OQL      0.356     395        6
#> 10 2002-2008 GXL   HVJ      0.081      90        6
#> # ... with 120 more rows
#> 
#> $interval_lvl
#> An object of class "Interval"
#> Slot "intervalData":
#> # A tibble: 4 x 4
#> # Groups:   Period [4]
#>   Period    PercentChange    St     U
#>   <fct>             <dbl> <dbl> <dbl>
#> 1 2012-2014         3.32  1.66   1.13
#> 2 2010-2012         4.23  2.12   1.13
#> 3 2008-2010         0.880 0.440  1.13
#> 4 2002-2008         5.18  0.864  1.13
#> 
#> 
#> $category_lvlGain
#> An object of class "Category"
#> Slot "lookupcolor":
#>       LEJ       GXL       IKW       KCI       PZV       ERH       HBS       WLE 
#> "#002F70" "#265BAB" "#6889D0" "#A3B4E5" "#D3DBF4" "#F6F6F6" "#F7D3D3" "#E6A4A4" 
#>       WNH       OQL       HVJ 
#> "#CB6F70" "#9D3D3D" "#5F1415" 
#> 
#> Slot "categoryData":
#> # A tibble: 23 x 6
#> # Groups:   Period, To [23]
#>    Period    To    Interval  GG_km2   Gtj    St
#>    <fct>     <fct>    <int>   <dbl> <dbl> <dbl>
#>  1 2012-2014 HVJ          2   0.195 0.337  1.66
#>  2 2012-2014 LEJ          2 612.    3.92   1.66
#>  3 2012-2014 PZV          2  14.9   0.510  1.66
#>  4 2012-2014 WLE          2   6.79  2.67   1.66
#>  5 2012-2014 WNH          2 110.    1.14   1.66
#>  6 2010-2012 LEJ          2 707.    4.84   2.12
#>  7 2010-2012 OQL          2   2.76  0.951  2.12
#>  8 2010-2012 PZV          2  47.0   1.18   2.12
#>  9 2010-2012 WLE          2   1.90  0.792  2.12
#> 10 2010-2012 WNH          2 189.    2.00   2.12
#> # ... with 13 more rows
#> 
#> Slot "categoryStationarity":
#> # A tibble: 12 x 5
#>    To     Gain     N Stationarity Test 
#>    <fct> <int> <int> <chr>        <chr>
#>  1 HBS       1     4 Active Gain  N    
#>  2 LEJ       2     4 Active Gain  N    
#>  3 OQL       2     4 Active Gain  N    
#>  4 PZV       2     4 Active Gain  N    
#>  5 WLE       2     4 Active Gain  N    
#>  6 WNH       1     4 Active Gain  N    
#>  7 HVJ       3     4 Dormant Gain N    
#>  8 LEJ       2     4 Dormant Gain N    
#>  9 OQL       1     4 Dormant Gain N    
#> 10 PZV       2     4 Dormant Gain N    
#> 11 WLE       2     4 Dormant Gain N    
#> 12 WNH       3     4 Dormant Gain N    
#> 
#> 
#> $category_lvlLoss
#> An object of class "Category"
#> Slot "lookupcolor":
#>       LEJ       GXL       IKW       KCI       PZV       ERH       HBS       WLE 
#> "#002F70" "#265BAB" "#6889D0" "#A3B4E5" "#D3DBF4" "#F6F6F6" "#F7D3D3" "#E6A4A4" 
#>       WNH       OQL       HVJ 
#> "#CB6F70" "#9D3D3D" "#5F1415" 
#> 
#> Slot "categoryData":
#> # A tibble: 29 x 6
#> # Groups:   Period, From [29]
#>    Period    From  Interval GL_km2     Lti    St
#>    <fct>     <fct>    <int>  <dbl>   <dbl> <dbl>
#>  1 2012-2014 ERH          2   3.49  0.224   1.66
#>  2 2012-2014 GXL          2  15.4   0.365   1.66
#>  3 2012-2014 IKW          2  25.3   0.845   1.66
#>  4 2012-2014 KCI          2  46.9   0.634   1.66
#>  5 2012-2014 LEJ          2 111.    0.759   1.66
#>  6 2012-2014 PZV          2 541.   13.6     1.66
#>  7 2012-2014 WNH          2   1.26  0.0133  1.66
#>  8 2010-2012 ERH          2   4.44  0.283   2.12
#>  9 2010-2012 GXL          2  11.1   0.263   2.12
#> 10 2010-2012 IKW          2  29.7   0.974   2.12
#> # ... with 19 more rows
#> 
#> Slot "categoryStationarity":
#> # A tibble: 14 x 5
#>    From   Loss     N Stationarity Test 
#>    <fct> <int> <int> <chr>        <chr>
#>  1 ERH       2     4 Active Loss  N    
#>  2 GXL       1     4 Active Loss  N    
#>  3 IKW       2     4 Active Loss  N    
#>  4 KCI       2     4 Active Loss  N    
#>  5 LEJ       1     4 Active Loss  N    
#>  6 OQL       1     4 Active Loss  N    
#>  7 PZV       3     4 Active Loss  N    
#>  8 ERH       2     4 Dormant Loss N    
#>  9 GXL       3     4 Dormant Loss N    
#> 10 IKW       2     4 Dormant Loss N    
#> 11 KCI       2     4 Dormant Loss N    
#> 12 LEJ       3     4 Dormant Loss N    
#> 13 PZV       1     4 Dormant Loss N    
#> 14 WNH       4     4 Dormant Loss Y    
#> 
#> 
#> $transition_lvlGain_n
#> An object of class "Transition"
#> Slot "lookupcolor":
#>       LEJ       GXL       IKW       KCI       PZV       ERH       HBS       WLE 
#> "#002F70" "#265BAB" "#6889D0" "#A3B4E5" "#D3DBF4" "#F6F6F6" "#F7D3D3" "#E6A4A4" 
#>       WNH       OQL       HVJ 
#> "#CB6F70" "#9D3D3D" "#5F1415" 
#> 
#> Slot "transitionData":

#> Warning: Factor `Period` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `From` contains implicit NA, consider using `forcats::fct_explicit_na`

#> # A tibble: 1 x 7
#> # Groups:   Period, From [1]
#>   Period From  To    Interval T_i2n_km2  Rtin   Wtn
#>   <fct>  <fct> <fct>    <int>     <dbl> <dbl> <dbl>
#> 1 <NA>   <NA>  <NA>        NA         0    NA    NA
#> 
#> Slot "transitionStationarity":
#> # A tibble: 2 x 5
#>   From   Loss     N Stationarity   Test 
#>   <fct> <int> <int> <chr>          <chr>
#> 1 <NA>      0     1 targeted by Ap N    
#> 2 <NA>      0     1 avoided by Ap  N    
#> 
#> 
#> $transition_lvlLoss_m
#> An object of class "Transition"
#> Slot "lookupcolor":
#>       LEJ       GXL       IKW       KCI       PZV       ERH       HBS       WLE 
#> "#002F70" "#265BAB" "#6889D0" "#A3B4E5" "#D3DBF4" "#F6F6F6" "#F7D3D3" "#E6A4A4" 
#>       WNH       OQL       HVJ 
#> "#CB6F70" "#9D3D3D" "#5F1415" 
#> 
#> Slot "transitionData":

#> Warning: Factor `Period` contains implicit NA, consider using `forcats::fct_explicit_na`

#> Warning: Factor `To` contains implicit NA, consider using `forcats::fct_explicit_na`

#> # A tibble: 1 x 7
#> # Groups:   Period, To [1]
#>   Period To    From  Interval T_m2j_km2  Qtmj   Vtm
#>   <fct>  <fct> <fct>    <int>     <dbl> <dbl> <dbl>
#> 1 <NA>   <NA>  <NA>        NA         0    NA    NA
#> 
#> Slot "transitionStationarity":
#> # A tibble: 2 x 5
#>   To     Gain     N Stationarity Test 
#>   <fct> <int> <int> <chr>        <chr>
#> 1 <NA>      0     1 targeted SG  N    
#> 2 <NA>      0     1 avoided SG   N    
#> 
#>