Methods for function plot in package OpenLand

Plot Intensity objects based on Intensity Analysis output.

plot(x, y, ...)

# S4 method for Interval,ANY
plot(
  x,
  y,
  labels = c(leftlabel = "Interval Change Area (percent of map)", rightlabel =
    "Annual Change Area (percent of map)"),
  title = NA,
  labs = c(type = "Changes", ur = "Uniform Intensity"),
  marginplot = c(lh = -10, rh = 0),
  leg_curv = c(x = 0.1, y = 0.1),
  color_bar = c(fast = "#B22222", slow = "#006400", area = "gray40"),
  fontsize_ui = 10,
  ...
)

# S4 method for Category,ANY
plot(
  x,
  y,
  labels = c(leftlabel = "Annual Change Area (km2 or pixels)", rightlabel =
    "Annual Change Intensity (percent of category)"),
  title = NA,
  labs = c(type = "Categories", ur = "Uniform Intensity"),
  marginplot = c(lh = 0.5, rh = 0.5),
  leg_curv = c(x = 0.1, y = 0.1),
  fontsize_ui = 10,
  ...
)

# S4 method for Transition,ANY
plot(
  x,
  y,
  labels = c(leftlabel = "Annual Transition Area (km2 or pixels)", rightlabel =
    "Annual Transition Intensity (percent of category)"),
  title = NA,
  labs = c(type = "Categories", ur = "Uniform Intensity"),
  marginplot = c(lh = 0.5, rh = 0.5),
  leg_curv = c(x = 0.1, y = 0.1),
  fontsize_ui = 10,
  ...
)

Arguments

x
An intensity object generated by intensityAnalysis.

y
ignored.

...
additional arguments for theme parameters from ggplot2, see theme.

labels
character. Left and right axis titles(caption).

title
character. Main title.

labs
character. The lateral legend.

marginplot
numeric. Adjustment of the origins of left and right part of the plots.

leg_curv
numeric. x and y values that control the arrow size and position pointing to the Uniform Intensity vertical line.

color_bar
character. Colors defined for the fast, slow and area bars (only for an Interval object).

fontsize_ui
numeric. Fontsize of the uniform intensity percent in the plot.

Interval
The class.

Category
The class.

Transition
The class.

Value

An intensity graph